fix(ios): build failure when using custom native marks
Build fails on importing `RNPeroformance.h` in Objective-C file, because `chrono` is a part of STL library which cannot be found in plain Objective-C files. The idea of fix is to prevent it to be included in a header.

`RNPerformanceEntryWasAddedNotification` declaration and initialization were split to prevent build failure as well.
